Implement Mean as an extension trait to Iterator
Since each trait 'instance' can only be implemented once for each type,
we could not implement `Mean` for both `I: Iterator<Item = f32>` and
`Item = &f32`, as the `I` could technically be the same.

This patch rewrites it to instead of taking an Input type parameter to
differentiate the traits, and takes an approach similar to `Sum` in the
standard library.

It also has the advantage that users could extend what can be taken a
'mean' of by implementing `Mean` themselves.

Technically, this could be made a generic impl for types that implement
a certain set of traits (sum + div), but that is potentially not super
useful and just makes things more complicated.

Signed-off-by: Marcel Müller <neikos@neikos.email>